refactor(cli): split localization commands
