+++ packages/cli/src/registry.ts
@@ -7,7 +7,7 @@ import {
 import { frameworkErrorCodes, invalidInvocationError } from "./runtime/errors.js";
 import { logLevels } from "./runtime/logger.js";
 import { createInstallCommand } from "./install/install-command.js";
-import { createLocalizationCommands } from "./localization/localization-command.js";
+import { createLocalizationCommands } from "./localization/index.js";
 
 export interface CommandOption {
   readonly longFlag: string;
